+++ CHANGELOG.md
@@ -13,6 +13,8 @@ CHANGELOG
   * `ONEY`
   * `POSCONNECT`
 * Added new type to the `Event.Type` enum: `PAYOUT_CHANGE`
+* Added support for new Device output:
+  * `/device/local_time`
 
 1.8.0 (2018-01-19)
 ------------------